# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel AGT",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Display a list of the transactions",
                "description": "get all transactions on database",
                "operationId": "716a962b580bc03cf3aa45dc2ab6c0e2",
                "responses": {
                    "200": {
                        "description": "List of tasks"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Store a new transaction",
                "description": "Store a new transaction on database",
                "operationId": "238affdbe033b73010cbace3ad94520e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "purchase_value": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "sale_value": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction stored"
                    }
                }
            }
        },
        "/transactions/{id}": {
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update a transaction",
                "description": "Update a transaction on database",
                "operationId": "bc5fceca8aaeaec95ec86f18ff8d067b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "purchase_value": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "sale_value": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction stored"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update a transaction",
                "description": "Update a transaction on database",
                "operationId": "7f382141ea50c7d87ab0f7b3bc03f34e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Transaction stored"
                    }
                }
            }
        }
    }
}